Move objects fixtures from local to fixtures pkg
