+++ internal/pkg/local/delete_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
+	"github.com/keboola/keboola-as-code/internal/pkg/fixtures"
 )
 
 func TestLocalDeleteModel(t *testing.T) {
@@ -13,7 +14,7 @@ func TestLocalDeleteModel(t *testing.T) {
 	manager, _ := newTestLocalManager(t)
 	fs := manager.fs
 
-	record := &MockedRecord{}
+	record := &fixtures.MockedRecord{}
 	assert.NoError(t, manager.manifest.PersistRecord(record))
 	_, found := manager.manifest.GetRecord(record.Key())
 	assert.True(t, found)
